Frontend: Company CRUD and service connection with api. (#11)
* feat: person relationships endpoints with proper service implementation and documentation

* feat: people service now calls the api relationship endpoint and component render the tabs conditionally based on the PersonRelationship data

* fix: format of accordion item header and icons fs-4

* fix: resolve duplicate key error on tax creation without setting Id

* feat: implement companies management feature with service, component, and routing setup

* feat: add truncate pipe

* git commit -m "refactor(company): enhance CompanyService and management features

- Refactored CompanyService to use constructor parameters directly
- Updated CreateAsync to add success messages and status codes
- Removed unused Student model import in CompanyConfiguration
- Renamed postalCode to zipCode in Person model for consistency
- Clarified API endpoint naming
- Added edit, delete, and detail icons (feat)
- Enhanced CompaniesService with create, update, delete methods
- Improved index-companies with modal triggers for CRUD actions
- Updated create-people, update-people to use zipCode
- Cleaned up view-people to display zipCode
- Added ActivitySector, CompanySize enums for type safety
- Created components for company CRUD operations (feat)
- Updated styles for consistent UI (style)

Closes #123"

* feat: enhance UI components with new icons and improve dropdown actions

* feat: implement delete company modal and enhance routing for companies

* feat: add company retrieval and update functionality with form validation

+++ NERBABO.Backend/NERBABO.ApiService/Core/Global/Models/Tax.cs
@@ -9,8 +9,8 @@ namespace NERBABO.ApiService.Core.Global.Models;
 public class Tax : Entity<int>
 {
     public string Name { get; set; } = string.Empty;
-    public int ValuePercent { get; set; }
-    public float ValueDecimal => (float)ValuePercent / 100;
+    public float ValuePercent { get; set; }
+    public float ValueDecimal => ValuePercent / 100;
     public bool IsActive { get; set; }
     public TaxEnum Type { get; set; } = TaxEnum.IVA;
 
@@ -21,7 +21,7 @@ public class Tax : Entity<int>
 
     // constructors
     public Tax() { }
-    public Tax(int id, string name, int valuePercent, TaxEnum type)
+    public Tax(int id, string name, float valuePercent, TaxEnum type)
     {
         Id = id;
         Name = name;
@@ -29,7 +29,7 @@ public class Tax : Entity<int>
         Type = type;
     }
 
-    public Tax(string name, int valuePercent, bool active, TaxEnum type)
+    public Tax(string name, float valuePercent, bool active, TaxEnum type)
     {
         Name = name;
         ValuePercent = valuePercent;
@@ -37,7 +37,7 @@ public class Tax : Entity<int>
         Type = type;
     }
 
-    public Tax(int id, string name, int valuePercent, bool active, TaxEnum type)
+    public Tax(int id, string name, float valuePercent, bool active, TaxEnum type)
     {
         Id = id;
         Name = name;
